backend/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-api-python-client==2.*
requests==2.*

